Traverse config.chunks in {min:1} order
This will allow us not to miss any new chunks created by splits.

Also we can keep track of the latest max, effeciently checking if
the current chunk is a subrange that we've already seen. I use
bsonWoCompare() to compare bson objects, so mongo shell from 2.5.3
or better is needed.