mdiag: #!/bin/bash, not #!/bin/sh
Since mdiag.sh includes bash-specific code (notably, use of the "<<-EOF"
indented here-document construct), it is not appropriate to #! a non-bash
Bourne shell (which, on some systems, may or may not have support for such
syntax, eg. when /bin/sh -> ash or dash).

Although the mdiag instructions usually say to run it via "sudo bash mdiag
...", this change makes work when chmodded +x and run directly (ie.  as "sudo
mdiag.sh ..."), on systems where /bin/sh is not bash.