mdiag: Fix fragility from using "<<-" construct.
The "<<-" here-document construct is a bash feature that will strip leading tab
characters when processing the following lines, including when looking for the
EOF marker.  Unfortunately, if the script has accidentally been mangled such
that the tab characters are converted to spaces, this means the script will
abort at this point with "unexpected end of file".  This causes the output to
be incomplete.

This patch fixes that problem by switching to a regular "<<" here-document, and
removing the indentation.  This comes at a small cost to the readability of
that section of code.  There are no longer any uses of "<<-", meaning that this
version of the script is no longer susceptible to this problem.